Weekend Commit.
Fixed image locations so not duplicate images in the repo.

Able to read and detect a change in the gcode files for Simplify3D and Repetier. the change will add a new print to the print stream with the print information. Will need to implement this stream and allow the user to choose which prints to add/drop. Will ned to test on desktop with slicing new models. Windows only, need to do testing to figure out where the files are located on Linux/Mac.

Able to minimize the program to the tray and open it back up. Will need to work on popup menus and everything else for this feature.